fix klockwork issue 51471
bitwise operation on different integer sizes. Fixed by
static_cast to larger integer size.

+++ util.h
@@ -91,7 +91,7 @@ static inline uint64_t hash(const std::string& str)
 
     for (auto i : str)
     {
-        ret ^= i;
+        ret ^= static_cast<uint64_t>(i);
         ret *= prime;
     }
